@@ -1,30 +1,40 @@
 {
-    "repositories": [{
-        "type": "composer",
-        "url": "https://packagist.org"
-    }],
-    "name": "wildbit/postmark-php",
-    "license": "MIT",
-    "description": "The officially supported client for Postmark (http://postmarkapp.com)",
-    "require": {
-        "php": ">=7.0.0",
-        "guzzlehttp/guzzle": "^6.0|^7.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^6.0.0"
-    },
-    "autoload": {
-        "psr-0": {
-            "Postmark\\": "src/"
-        }
-    },
-    "autoload-dev": {
-        "classmap": [
-            "src/",
-            "tests/"
-        ]
-    },
-    "scripts": {
-        "test": "phpunit"
+  "repositories": [
+    {
+      "type": "composer",
+      "url": "https://packagist.org"
     }
+  ],
+  "name": "wildbit/postmark-php",
+  "license": "MIT",
+  "description": "The officially supported client for Postmark (http://postmarkapp.com)",
+  "require": {
+    "php": "~8.0 || ~8.1 || ~8.2",
+    "guzzlehttp/guzzle": "^7.8"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^9",
+    "phpstan/phpstan": "^1.10",
+    "friendsofphp/php-cs-fixer": "^3.40"
+  },
+  "autoload": {
+    "psr-0": {
+      "Postmark\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "classmap": [
+      "src/",
+      "tests/"
+    ]
+  },
+  "scripts": {
+    "test": "phpunit"
+  },
+  "config": {
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "php-http/discovery": false
+    }
+  }
 }
