@@ -1,41 +1,44 @@
 <?php
+
 namespace Postmark\Models\Webhooks;
+
+use JsonSerializable;
 
 /**
  * Settings for Open webhooks.
  */
-class WebhookConfigurationOpenTrigger implements \JsonSerializable {
-
-    private $enabled;
-    private $postFirstOpenOnly;
+class WebhookConfigurationOpenTrigger implements JsonSerializable
+{
+    private bool $Enabled;
+    private bool $PostFirstOpenOnly;
 
     /**
      * Create a new WebhookConfigurationOpenTrigger.
      *
-     * @param boolean $enabled Specifies whether or not webhooks will be triggered by Open events.
-     * @param boolean $postFirstOpenOnly If enabled, Open webhooks will only POST on first open.
+     * @param bool $enabled           specifies whether or not webhooks will be triggered by Open events
+     * @param bool $postFirstOpenOnly if enabled, Open webhooks will only POST on first open
      */
-    public function __construct($enabled = false, $postFirstOpenOnly = false) {
-        $this->enabled = $enabled;
-        $this->postFirstOpenOnly = $postFirstOpenOnly;
+    public function __construct(bool $enabled, bool $postFirstOpenOnly)
+    {
+        $this->Enabled = $enabled;
+        $this->PostFirstOpenOnly = $postFirstOpenOnly;
     }
 
-    public function jsonSerialize() {
-        $retval = array(
-            "Enabled" => $this->enabled,
-            "PostFirstOpenOnly" => $this->postFirstOpenOnly
-        );
-
-        return $retval;
+    public function jsonSerialize(): array
+    {
+        return [
+            'Enabled' => $this->Enabled,
+            'PostFirstOpenOnly' => $this->PostFirstOpenOnly,
+        ];
     }
 
-    public function getEnabled() {
-        return $this->enabled;
+    public function getEnabled(): bool
+    {
+        return $this->Enabled;
     }
 
-    public function getPostFirstOpenOnly() {
-        return $this->postFirstOpenOnly;
+    public function getPostFirstOpenOnly(): bool
+    {
+        return $this->PostFirstOpenOnly;
     }
 }
-
-?>
