@@ -1,11 +1,15 @@
 <?php
 
+/** @noinspection PhpMissingFieldTypeInspection */
+
 namespace Postmark\Models;
+
+use Exception;
 
 /**
  * The exception thrown when the Postmark Client recieves an error from the API.
  */
-class PostmarkException extends \Exception {
+class PostmarkException extends Exception {
 	var $message;
 	var $httpStatusCode;
 	var $postmarkApiErrorCode;
