@@ -1,41 +1,44 @@
 <?php
+
 namespace Postmark\Models\Webhooks;
+
+use JsonSerializable;
 
 /**
  * Model describing Basic HTTP Authentication.
  */
-class HttpAuth implements \JsonSerializable {
-
+class HttpAuth implements JsonSerializable
+{
     private $username;
     private $password;
 
     /**
      * Create a new HttpAuth.
      *
-     * @param string $username Username to use.
-     * @param string $password Password to use.
+     * @param string $username username to use
+     * @param string $password password to use
      */
-    public function __construct($username = null, $password = null) {
+    public function __construct($username = null, $password = null)
+    {
         $this->username = $username;
         $this->password = $password;
     }
 
-    public function jsonSerialize() {
-        $retval = array(
-            "Username" => $this->username,
-            "Password" => $this->password
-        );
-
-        return $retval;
+    public function jsonSerialize(): array
+    {
+        return [
+            'Username' => $this->username,
+            'Password' => $this->password,
+        ];
     }
 
-    public function getUsername() {
+    public function getUsername(): ?string
+    {
         return $this->username;
     }
 
-    public function getPassword() {
+    public function getPassword(): ?string
+    {
         return $this->password;
     }
 }
-
-?>
