@@ -1,41 +1,41 @@
 <?php
+
 namespace Postmark\Models\Webhooks;
 
-/**
- * Settings for Bounce webhooks.
- */
-class WebhookConfigurationBounceTrigger implements \JsonSerializable {
+use JsonSerializable;
 
-    private $enabled;
-    private $includeContent;
+class WebhookConfigurationBounceTrigger implements JsonSerializable
+{
+    private $Enabled;
+    private $IncludeContent;
 
     /**
      * Create a new WebhookConfigurationBounceTrigger.
      *
-     * @param boolean $enabled Specifies whether or not webhooks will be triggered by Bounce events.
-     * @param boolean $includeContent Specifies whether or not the full content of the email bounce is included in webhook POST.
+     * @param bool $enabled        specifies whether the webhooks will be triggered by Bounce events
+     * @param bool $includeContent specifies whether the full content of the email bounce is included in webhook POST
      */
-    public function __construct($enabled = false, $includeContent = false) {
-        $this->enabled = $enabled;
-        $this->includeContent = $includeContent;
+    public function __construct(bool $enabled, bool $includeContent)
+    {
+        $this->Enabled = $enabled;
+        $this->IncludeContent = $includeContent;
     }
 
-    public function jsonSerialize() {
-        $retval = array(
-            "Enabled" => $this->enabled,
-            "IncludeContent" => $this->includeContent
-        );
-
-        return $retval;
+    public function jsonSerialize(): array
+    {
+        return [
+            'Enabled' => $this->Enabled,
+            'IncludeContent' => $this->IncludeContent,
+        ];
     }
 
-    public function getEnabled() {
-        return $this->enabled;
+    public function getEnabled(): bool
+    {
+        return $this->Enabled;
     }
 
-    public function getIncludeContent() {
-        return $this->includeContent;
+    public function getIncludeContent(): bool
+    {
+        return $this->IncludeContent;
     }
 }
-
-?>
