@@ -1,33 +1,35 @@
 <?php
+
 namespace Postmark\Models\Webhooks;
+
+use JsonSerializable;
 
 /**
  * Settings for Delivery webhooks.
  */
-class WebhookConfigurationDeliveryTrigger implements \JsonSerializable {
-
+class WebhookConfigurationDeliveryTrigger implements JsonSerializable
+{
     private $enabled;
 
     /**
      * Create a new WebhookConfigurationDeliveryTrigger.
      *
-     * @param boolean $enabled Specifies whether or not webhooks will be triggered by Delivery events.
+     * @param bool $enabled specifies whether or not webhooks will be triggered by Delivery events
      */
-    public function __construct($enabled = false) {
+    public function __construct($enabled = false)
+    {
         $this->enabled = $enabled;
     }
 
-    public function jsonSerialize() {
-        $retval = array(
-            "Enabled" => $this->enabled
-        );
-
-        return $retval;
+    public function jsonSerialize(): array
+    {
+        return [
+            'Enabled' => $this->enabled,
+        ];
     }
 
-    public function getEnabled() {
+    public function getEnabled(): bool
+    {
         return $this->enabled;
     }
 }
-
-?>
