@@ -2,128 +2,125 @@
 
 namespace Postmark\Tests;
 
-require_once __DIR__ . "/PostmarkClientBaseTest.php";
+require_once __DIR__ . '/PostmarkClientBaseTest.php';
 
-use Postmark\PostmarkAdminClient as PostmarkAdminClient;
+use Postmark\PostmarkAdminClient;
 
-class PostmarkAdminClientSenderSignatureTest extends PostmarkClientBaseTest {
+/**
+ * @internal
+ *
+ * @coversNothing
+ */
+class PostmarkAdminClientSenderSignatureTest extends PostmarkClientBaseTest
+{
+    public static function tearDownAfterClass(): void
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	static function tearDownAfterClass() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $signatures = $client->listSenderSignatures();
 
-		$signatures = $client->listSenderSignatures();
+        foreach ($signatures->getSenderSignatures() as $key => $value) {
+            if (preg_match('/test-php.+/', $value->getName()) > 0) {
+                $client->deleteSenderSignature($value->getID());
+            }
+        }
+    }
 
-		foreach ($signatures->senderSignatures as $key => $value) {
-			if (preg_match('/test-php.+/', $value->name) > 0) {
-				$client->deleteSenderSignature($value->id);
-			}
-		}
-	}
+    public function testClientCanGetSignatureList()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanGetSignatureList() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $sigs = $client->listSenderSignatures();
 
-		$sigs = $client->listSenderSignatures();
+        $this->assertGreaterThan(0, $sigs->getTotalCount());
+        $this->assertNotEmpty($sigs->getSenderSignatures());
+    }
 
-		$this->assertGreaterThan(0, $sigs->totalCount);
-		$this->assertNotEmpty($sigs->senderSignatures);
-	}
+    public function testClientCanGetSingleSignature()
+    {
+        $tk = parent::$testKeys;
 
-	function testClientCanGetSingleSignature() {
-		$tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $id = $client->listSenderSignatures()->getSenderSignatures()[0]->getID();
+        $sig = $client->getSenderSignature($id);
 
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
-		$id = $client->listSenderSignatures()->senderSignatures[0]->id;
-		$sig = $client->getSenderSignature($id);
+        $this->assertNotEmpty($sig->getName());
+    }
 
-		$this->assertNotEmpty($sig->name);
-	}
+    public function testClientCanCreateSignature()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanCreateSignature() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
+        $sender = str_replace('[TOKEN]', 'test-php-create' . date('U'), $i);
 
-		$i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
-		$sender = str_replace('[TOKEN]', 'test-php-create' . date('U'), $i);
+        $sig = $client->createSenderSignature($sender, 'test-php-create-' . date('U'));
 
-		$sig = $client->createSenderSignature($sender, 'test-php-create-' . date('U'));
+        $this->assertNotEmpty($sig->getID());
+    }
 
-		$this->assertNotEmpty($sig->id);
-	}
+    public function testClientCanEditSignature()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanEditSignature() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $name = 'test-php-edit-' . date('U');
 
-		$name = 'test-php-edit-' . date('U');
-
-		$i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
-		$sender = str_replace('[TOKEN]', 'test-php-edit' . date('U'), $i);
+        $i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
+        $sender = str_replace('[TOKEN]', 'test-php-edit' . date('U'), $i);
 
         $exploded = explode('@', $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE);
-		$returnPath = 'test.' . $exploded[1];
+        $returnPath = 'test.' . $exploded[1];
 
-		$sig = $client->createSenderSignature($sender, $name, NULL, $returnPath);
+        $sig = $client->createSenderSignature($sender, $name, null, $returnPath);
 
-		$updated = $client->editSenderSignature(
-			$sig->id, $name . '-updated', NULL, 'updated-' . $returnPath);
+        $updated = $client->editSenderSignature(
+            $sig->getID(),
+            $name . '-updated',
+            null,
+            'updated-' . $returnPath
+        );
 
-		$this->assertNotSame($sig->name, $updated->name);
-		$this->assertNotSame($sig->returnpathdomain, $updated->returnpathdomain);
-	}
+        $this->assertNotSame($sig->getName(), $updated->getName());
+        $this->assertNotSame($sig->getReturnpathdomain(), $updated->getReturnpathdomain());
+    }
 
-	function testClientCanDeleteSignature() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+    public function testClientCanDeleteSignature()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-		$i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
-		$sender = str_replace('[TOKEN]', 'test-php-delete' . date('U'), $i);
+        $i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
+        $sender = str_replace('[TOKEN]', 'test-php-delete' . date('U'), $i);
 
-		$name = 'test-php-delete-' . date('U');
-		$sig = $client->createSenderSignature($sender, $name);
+        $name = 'test-php-delete-' . date('U');
+        $sig = $client->createSenderSignature($sender, $name);
 
-		$client->deleteSenderSignature($sig->id);
+        $client->deleteSenderSignature($sig->getID());
 
-		$sigs = $client->listSenderSignatures()->senderSignatures;
+        $sigs = $client->listSenderSignatures()->getSenderSignatures();
 
-		foreach ($sigs as $key => $value) {
-			$this->assertNotSame($sig->name, $value->name);
-		}
+        foreach ($sigs as $key => $value) {
+            $this->assertNotSame($sig->getName(), $value->getName());
+        }
+    }
 
-	}
+    public function testClientCanRequestNewVerificationForSignature()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanRequestNewVerificationForSignature() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
+        $sender = str_replace('[TOKEN]', 'test-php-reverify' . date('U'), $i);
 
-		$i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
-		$sender = str_replace('[TOKEN]', 'test-php-reverify' . date('U'), $i);
+        $name = 'test-php-reverify-' . date('U');
+        $sig = $client->createSenderSignature($sender, $name);
 
-		$name = 'test-php-reverify-' . date('U');
-		$sig = $client->createSenderSignature($sender, $name);
+        $result = $client->resendSenderSignatureConfirmation($sig->getID());
 
-		$result = $client->resendSenderSignatureConfirmation($sig->id);
-
-		$this->assertEquals(0, $result->ErrorCode);
-	}
-
-	function testClientCanVerifySPFForSignature() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
-
-		$name = 'test-php-spf-' . date('U');
-		$i = $tk->WRITE_TEST_SENDER_SIGNATURE_PROTOTYPE;
-
-		$sender = str_replace('[TOKEN]', 'test-php-spf-' . date('U'), $i);
-
-		$sig = $client->createSenderSignature($sender, $name);
-		$result = $client->verifySenderSignatureSPF($sig->id);
-
-		$this->assertTrue($result->SPFVerified);
-	}
-
+        $this->assertEquals(0, $result->getErrorCode());
+    }
 }
-
-?>
