@@ -2,82 +2,91 @@
 
 namespace Postmark\Tests;
 
-require_once __DIR__ . "/PostmarkClientBaseTest.php";
+require_once __DIR__ . '/PostmarkClientBaseTest.php';
 
-use Postmark\PostmarkAdminClient as PostmarkAdminClient;
+use Postmark\PostmarkAdminClient;
 
-class PostmarkAdminClientServersTest extends PostmarkClientBaseTest {
+/**
+ * @internal
+ *
+ * @coversNothing
+ */
+class PostmarkAdminClientServersTest extends PostmarkClientBaseTest
+{
+    public static function tearDownAfterClass(): void
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	static function tearDownAfterClass() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $servers = $client->listServers();
 
-		$servers = $client->listServers();
+        foreach ($servers->getServers() as $key => $value) {
+            if (preg_match('/^test-php.+/', $value->getName()) > 0 && !empty($value->getID())) {
+                $client->deleteServer($value->getID());
+            }
+        }
+    }
 
-		foreach ($servers->servers as $key => $value) {
-			if (preg_match('/^test-php.+/', $value->name) > 0) {
-				$client->deleteServer($value->id);
-			}
-		}
-	}
+    public function testClientCanGetServerList()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanGetServerList() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $servers = $client->listServers();
 
-		$servers = $client->listServers();
+        $this->assertGreaterThan(0, $servers->getTotalcount());
+        $this->assertNotEmpty($servers->getServers()[0]);
+    }
 
-		$this->assertGreaterThan(0, $servers['totalcount']);
-		$this->assertNotEmpty($servers->servers[0]);
-	}
+    public function testClientCanGetSingleServer()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanGetSingleServer() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $servers = $client->listServers();
+        $targetId = $servers->getServers()[0]->getID();
 
-		$servers = $client->listServers();
-		$targetId = $servers->servers[0]->id;
+        $server = $client->getServer($targetId);
 
-		$server = $client->getServer($targetId);
+        $this->assertNotEmpty($server->getName());
+    }
 
-		$this->assertNotEmpty($server->name);
-	}
+    public function testClientCanCreateServer()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanCreateServer() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $server = $client->createServer('test-php-create-' . date('c'));
+        $this->assertNotEmpty($server);
+    }
 
-		$server = $client->createServer('test-php-create-' . date('c'));
-		$this->assertNotEmpty($server);
-	}
+    public function testClientCanEditServer()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanEditServer() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $server = $client->createServer('test-php-edit-' . date('c'), 'purple');
 
-		$server = $client->createServer('test-php-edit-' . date('c'), 'purple');
+        $updateName = 'test-php-edit-' . date('c') . '-updated';
+        $serverUpdated = $client->editServer($server->getID(), $updateName, 'green');
 
-		$updateName = 'test-php-edit-' . date('c') . '-updated';
-		$serverUpdated = $client->editServer($server->id, $updateName, 'green');
+        $this->assertNotEmpty($serverUpdated);
+        $this->assertNotSame($serverUpdated->getName(), $server->getName());
+        $this->assertNotSame($serverUpdated->getColor(), $server->getColor());
+    }
 
-		$this->assertNotEmpty($serverUpdated);
-		$this->assertNotSame($serverUpdated->name, $server->name);
-		$this->assertNotSame($serverUpdated->color, $server->color);
-	}
+    public function testClientCanDeleteServer()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanDeleteServer() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkAdminClient($tk->WRITE_ACCOUNT_TOKEN, $tk->TEST_TIMEOUT);
+        $server = $client->createServer('test-php-delete-' . date('c'));
 
-		$server = $client->createServer('test-php-delete-' . date('c'));
+        $client->deleteServer($server->getID());
 
-		$client->deleteServer($server->id);
-
-		$serverList = $client->listServers();
-		foreach ($serverList->servers as $key => $value) {
-			$this->assertNotSame($value->id, $server->id);
-		}
-	}
+        $serverList = $client->listServers();
+        foreach ($serverList->getServers() as $key => $value) {
+            $this->assertNotSame($value->getID(), $server->getID());
+        }
+    }
 }
-
-?>
