@@ -2,21 +2,21 @@
 
 namespace Postmark\Tests;
 
-require_once __DIR__ . "/../vendor/autoload.php";
-require_once __DIR__ . "/TestingKeys.php";
+require_once __DIR__ . '/../vendor/autoload.php';
 
-use Postmark\PostmarkClientBase as PostmarkClientBase;
+require_once __DIR__ . '/TestingKeys.php';
 
-abstract class PostmarkClientBaseTest extends \PHPUnit\Framework\TestCase {
+use Postmark\PostmarkClientBase;
 
-	public static $testKeys;
+abstract class PostmarkClientBaseTest extends \PHPUnit\Framework\TestCase
+{
+    public static $testKeys;
 
-	public static function setUpBeforeClass() {
-		//get the config keys for the various tests
-		self::$testKeys = new TestingKeys();
-		PostmarkClientBase::$BASE_URL = self::$testKeys->BASE_URL ?: 'https://api.postmarkapp.com';
-		date_default_timezone_set("UTC");
-	}
+    public static function setUpBeforeClass(): void
+    {
+        // get the config keys for the various tests
+        self::$testKeys = new TestingKeys();
+        PostmarkClientBase::$BASE_URL = self::$testKeys->BASE_URL ?: 'https://api.postmarkapp.com';
+        date_default_timezone_set('UTC');
+    }
 }
-
-?>
