@@ -2,42 +2,49 @@
 
 namespace Postmark\Tests;
 
-require_once __DIR__ . "/PostmarkClientBaseTest.php";
+require_once __DIR__ . '/PostmarkClientBaseTest.php';
 
-use \Postmark\PostmarkClient;
+use Postmark\PostmarkClient;
 
-class PostmarkClientOutboundMessageTest extends PostmarkClientBaseTest {
+/**
+ * @internal
+ *
+ * @coversNothing
+ */
+class PostmarkClientOutboundMessageTest extends PostmarkClientBaseTest
+{
+    public function testClientCanSearchOutboundMessages()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkClient($tk->READ_SELENIUM_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanSearchOutboundMessages() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkClient($tk->READ_SELENIUM_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
+        $messages = $client->getOutboundMessages(10);
+        $this->assertNotEmpty($messages);
+        $this->assertCount(10, $messages->getMessages());
+    }
 
-		$messages = $client->getOutboundMessages(10);
-		$this->assertNotEmpty($messages);
-		$this->assertCount(10, $messages->Messages);
-	}
+    public function testClientCanGetOutboundMessageDetails()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkClient($tk->READ_SELENIUM_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanGetOutboundMessageDetails() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkClient($tk->READ_SELENIUM_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
+        $retrievedMessages = $client->getOutboundMessages(1, 50);
 
-		$retrievedMessages = $client->getOutboundMessages(1, 50);
+        $baseMessageId = $retrievedMessages->getMessages()[0]->getMessageID();
+        $message = $client->getOutboundMessageDetails($baseMessageId);
 
-		$baseMessageId = $retrievedMessages->Messages[0]["MessageID"];
-		$message = $client->getOutboundMessageDetails($baseMessageId);
+        $this->assertNotEmpty($message);
+    }
 
-		$this->assertNotEmpty($message);
-	}
+    public function testClientCanGetOutboundMessageDump()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkClient($tk->READ_SELENIUM_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanGetOutboundMessageDump() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkClient($tk->READ_SELENIUM_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
+        $retrievedMessages = $client->getOutboundMessages(1, 50);
+        $baseMessageId = $retrievedMessages->getMessages()[0]->getMessageID();
+        $message = $client->getOutboundMessageDump($baseMessageId);
 
-		$retrievedMessages = $client->getOutboundMessages(1, 50);
-		$baseMessageId = $retrievedMessages->Messages[0]["MessageID"];
-		$message = $client->getOutboundMessageDump($baseMessageId);
-
-		$this->assertNotEmpty($message);
-	}
+        $this->assertNotEmpty($message);
+    }
 }
-?>
