@@ -2,33 +2,38 @@
 
 namespace Postmark\Tests;
 
-require_once __DIR__ . "/PostmarkClientBaseTest.php";
+require_once __DIR__ . '/PostmarkClientBaseTest.php';
 
-use \Postmark\PostmarkClient;
+use Postmark\PostmarkClient;
 
-class PostmarkClientRuleTriggerTest extends PostmarkClientBaseTest {
+/**
+ * @internal
+ *
+ * @coversNothing
+ */
+class PostmarkClientRuleTriggerTest extends PostmarkClientBaseTest
+{
+    // todo - teardowns to clean up rules that were created and not deleted.
 
-	//todo - teardowns to clean up rules that were created and not deleted.
+    public function testClientCanGetRuleTriggers()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkClient($tk->WRITE_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanGetRuleTriggers() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkClient($tk->WRITE_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
+        $triggers = $client->listInboundRuleTriggers();
 
-		$triggers = $client->listInboundRuleTriggers();
+        $this->assertNotEmpty($triggers);
+    }
 
-		$this->assertNotEmpty($triggers);
-	}
+    public function testClientCanCreateAndDeleteRuleTriggers()
+    {
+        $tk = parent::$testKeys;
+        $client = new PostmarkClient($tk->WRITE_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
 
-	function testClientCanCreateAndDeleteRuleTriggers() {
-		$tk = parent::$testKeys;
-		$client = new PostmarkClient($tk->WRITE_TEST_SERVER_TOKEN, $tk->TEST_TIMEOUT);
+        $trigger = $client->createInboundRuleTrigger('test.php+' . uniqid('', true) . '@example.com');
+        $this->assertNotEmpty($trigger);
 
-		$trigger = $client->createInboundRuleTrigger('test.php+' . uniqid("", true) . '@example.com');
-		$this->assertNotEmpty($trigger);
-
-		$client->deleteInboundRuleTrigger($trigger->ID);
-		//Not throwing an exception here constitutes passing.
-	}
+        $client->deleteInboundRuleTrigger($trigger->getID());
+        // Not throwing an exception here constitutes passing.
+    }
 }
-
-?>
